refactoring--added some folders and moved components into them
