+++ src/components/about/mobile/CloseButtonMobile.tsx
@@ -1,16 +1,16 @@
 import React from 'react'
 import {styled, Typography, Grid, ButtonBase} from '@material-ui/core'
-import { theme } from "../theme";
+import { theme } from "../../../theme";
 
 
 const StyleGrid = styled(Grid)({
   margin: 'auto',
   width: '100%',
-  height: '4.5vw',
-  padding: '1.5vw 1.5vw 1.5vw 2.5vw',
+  height: '4.5rem',
+  padding: '1.5rem 1.5rem 1.5rem 2.5rem',
   background: 'transparent',
   boxSizing: 'border-box',
-  boxShadow: '0 0.1875vw 0.375vw 0 rgba(0, 0, 0, 0.16), inset 0 0.1875vw 0.375vw 0 rgba(0, 0, 0, 0.16)',
+  boxShadow: '0 0.1875rem 0.375rem 0 rgba(0, 0, 0, 0.16), inset 0 0.1875rem 0.375rem 0 rgba(0, 0, 0, 0.16)',
   borderRadius: 0,
   border: 'solid 2px ' + theme.palette.text.primary,
   '&:hover': {
@@ -21,11 +21,11 @@ const StyleGrid = styled(Grid)({
 
 const StyleText = styled(Typography)({
   width: '100%',
-  height: '1vw',
+  height: '1rem',
   margin: 'auto',
   padding: 0,
   fontFamily: theme.typography.fontFamily,
-  fontSize: '0.938vw',
+  fontSize: '0.938rem',
   fontWeight: 'bold',
   fontStretch: "normal",
   fontStyle: "normal",
@@ -36,9 +36,9 @@ const StyleText = styled(Typography)({
 });
 
 const StyleArrow = styled('img')({
-  width: '1.5vw',
-  height: '1.5vw',
-  margin: '0 0 0 1.75vw',
+  width: '1.5rem',
+  height: '1.5rem',
+  margin: '0 0 0 1.75rem',
   padding: 0,
   objectFit: "contain"
 });
@@ -48,7 +48,7 @@ interface Props {
   handleClick: () => void
 }
 
-export const CloseButton: React.FC<Props> = (props: Props) => {
+export const CloseButtonMobile: React.FC<Props> = (props: Props) => {
 
   return (
     <ButtonBase onClick={() => props.handleClick()}>
