+++ src/components/about/mobile/PitchTitleBoxMobile.tsx
@@ -1,64 +1,53 @@
 import React from 'react'
 import {Grid, styled, Typography,} from '@material-ui/core'
-import { theme } from "../theme";
+import { theme } from "../../../theme";
 
 
 const StyledGrid = styled(Grid)({
   margin: 'auto',
   width: '100%',
-  height: '16.2vw',
+  height: '9.375rem',
   background: 'transparent',
   boxSizing: 'border-box',
   position: 'relative'
 });
 
 const StyledText = styled(Typography)({
-  marginLeft: '1.8vw',
+  maxWidth: '19.438rem',
+  marginLeft: '1.438rem',
   fontFamily: theme.typography.fontFamily,
-  fontSize: '3.25vw',
+  fontSize: '1.75rem',
   fontWeight: 600,
   fontStretch: "normal",
   fontStyle: "normal",
-  lineHeight: 1,
-  letterSpacing: '-1.6px',
+  lineHeight: 1.5,
+  letterSpacing: '-0.84px',
   textAlign: "left",
   color: theme.palette.text.primary
 });
 
 const StyledRings = styled('img')({
-  width: '19.125vw',
-  height: '19.063vw',
+  width: '12.191rem',
+  height: '12.334rem',
   objectFit: 'contain',
   position: 'absolute',
-  top: 0,
+  top: -3,
   right: 0,
   zIndex: 1
 });
 
-const Accent = styled('div')({
-  height: '5vw',
-  position: 'absolute',
-  top: 0,
-  left: '50%',
-  right: '50%',
-  zIndex: 2
-});
-
-
 interface Props {
   text: string;
   classes?: string;
-  accentStyle?: string;
 }
 
-export const PitchTitleBox: React.FC<Props> = (props: Props) => {
+export const PitchTitleBoxMobile: React.FC<Props> = (props: Props) => {
   return (
     <StyledGrid className={props.classes} container alignItems='center'>
       <Grid item>
         <StyledText>{props.text}</StyledText>
       </Grid>
       <StyledRings src='imgs/concentric-rings-left.svg' alt={'concentric rings flourish'} />
-      <Accent style={{borderRight: props.accentStyle}} />
     </StyledGrid>
   );
 }
